Read attributes from Tileserver; communicate to carto-style-renderer using a POST.
We modified the format of the payload that tileserver
sends to carto-renderer to pass the attributes associated
with each feature. We're unpacking that now, attaching
it to Mapnik Features, and using it to build the
Mapnik-rendered tile.

The payload communication between carto-style-renderer
and carto-renderer became a larger bottleneck for us
due to payload size. We added POST requests to
carto-style-renderer and use that interface here.